requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
decorator==5.1.1
emcee==3.1.4
fastapi==0.110.0
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2024.2.0
GPy==1.13.1
gpytorch==1.11
grpclib==0.4.7
gstools==1.5.1
h2==4.1.0
hankel==1.2.2
hpack==4.0.0
hyperframe==6.0.1
idna==3.6
imageio==2.33.1
jaxtyping==0.2.28
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
linear-operator==0.5.2
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
meshio==5.3.5
modAL-python==0.4.2.1
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numba==0.59.0
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==2.2.0
paramz==0.9.6
pillow==10.2.0
protobuf==4.25.3
pyaml==23.12.0
pyasn1==0.5.1
pydantic==2.6.3
pydantic_core==2.16.3
pyDOE2==1.3.0
pyevtk==1.6.0
Pygments==2.17.2
PyKrige==1.7.1
pyparsing==3.1.1
PySimpleGUI==5.0.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
rich==13.7.0
rsa==4.9
scikit-gstat==1.0.16
scikit-learn==1.4.0
scikit-optimize==0.9.0
scipy==1.11.4
sigtools==4.0.1
six==1.16.0
skorch==0.9.0
sniffio==1.3.1
starlette==0.36.3
sympy==1.12
synchronicity==0.6.2
tabulate==0.9.0
tblib==3.0.0
threadpoolctl==3.2.0
toml==0.10.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.1
triton==2.2.0
typeguard==2.13.3
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.7
typing_extensions==4.10.0
tzdata==2023.4
watchfiles==0.21.0
yarl==1.9.4
